refactor: SectionAddAcceptanceTest (Section 인수 테스트) 관련 코드 리팩터링
