+++ src/test/java/nextstep/subway/line/SectionAddAcceptanceTestRequest.java
@@ -5,16 +5,16 @@ import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
 import nextstep.subway.line.dto.LineRequest;
 import nextstep.subway.line.dto.SectionRequest;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 
 import java.util.HashMap;
 import java.util.Map;
 
-import static org.assertj.core.api.Assertions.assertThat;
+import static nextstep.subway.line.LineTestCommon.createLineParams;
+import static nextstep.subway.line.LineTestCommon.createResponse;
 
-public class SectionAddAcceptanceTestHelper {
 
+public class SectionAddAcceptanceTestRequest {
     public static ExtractableResponse<Response> 노선에_지하철역_등록_요청(Long lineId, Long upStationId, Long downStationId, long distance) {
         SectionRequest sectionRequest = new SectionRequest(upStationId, downStationId, distance);
 
@@ -28,17 +28,12 @@ public class SectionAddAcceptanceTestHelper {
                 extract();
     }
 
-    public static void 지하철_노선에_지하철역_등록됨(ExtractableResponse<Response> response) {
-        assertThat(response.statusCode()).isEqualTo(HttpStatus.CREATED.value());
-        assertThat(response.header("Location")).isNotBlank();
-    }
-
     public static LineRequest 노선_구간_요청_생성(String lineName, String lineColor, Long upStationId, Long downStationId, long distance) {
-        return LineTestCommon.createLineParams(lineName, lineColor, upStationId, downStationId, distance);
+        return createLineParams(lineName, lineColor, upStationId, downStationId, distance);
     }
 
     public static ExtractableResponse<Response> 노선_등록시_구간_등록되어_있음(LineRequest request) {
-        return LineTestCommon.createResponse(request, "/lines");
+        return createResponse(request, "/lines");
     }
 
     public static ExtractableResponse<Response> 지하철역_등록되어_있음(String name) {
@@ -57,9 +52,5 @@ public class SectionAddAcceptanceTestHelper {
                 then().
                 log().all().
                 extract();
-    }
-
-    public static void 지하철_노선_구간_등록_예외_케이스로_등록되지_않음(ExtractableResponse<Response> response) {
-        assertThat(response.statusCode()).isEqualTo(HttpStatus.BAD_REQUEST.value());
     }
 }
